Fixed up the AllowedByBoth type
